feat: move more code from MapView (and convert to TS)
