+++ src/components/map/index.ts
@@ -1,5 +1,4 @@
 export { default as DrawingToolbar } from './DrawingToolbar';
 export { default as MapContextMenu } from './MapContextMenu';
 export { default as MapControls } from './MapControls';
-export { default as MapRotationTextBox } from './MapRotationTextBox';
-export { default as MapToolbar } from './MapToolbar';
+export { default as MapToolbars } from './MapToolbars';
